# packages/server/docs/swagger.yaml
definitions:
  models.ActivityRequest:
    properties:
      activity:
        type: string
    type: object
  models.CreateDatasetBranchRequest:
    properties:
      branch_name:
        type: string
    type: object
  models.CreateDatasetRequest:
    properties:
      branch_names:
        items:
          type: string
        type: array
      is_public:
        type: boolean
      readme:
        $ref: '#/definitions/models.ReadmeRequest'
      wiki:
        type: string
    type: object
...
  models.HashRequest:
    properties:
      hash:
        type: string
    type: object
  models.LogRequest:
    properties:
      data:
        type: string
      key:
        type: string
    type: object
...
  models.ReadmeRequest:
    properties:
      content:
        type: string
      file_type:
        type: string
    type: object
...
info:
  contact:
    email: contact@pureml.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /org/:
    get:
      consumes:
      - '*/*'
      description: Get all user organizations.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all user organizations.
      tags:
      - Organization
  /org/{orgId}/add:
    post:
      consumes:
      - '*/*'
      description: Add a user to an organization. Only accessible by owners of the
        organization.
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a user to an organization.
      tags:
      - Organization
  /org/{orgId}/dataset/{datasetName}:
    get:
      consumes:
      - '*/*'
      description: Get specific dataset of an organization
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get specific dataset of an organization
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/activity/{category}:
    get:
      consumes:
      - '*/*'
      description: Get activity of a dataset for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get activity of a dataset for a category
      tags:
      - Dataset
    post:
      consumes:
      - '*/*'
      description: Add activity of a dataset for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Activity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add activity of a dataset for a category
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/activity/{category}/{activityUUID}:
    post:
      consumes:
      - '*/*'
      description: Update activity of a dataset for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Activity UUID
        in: path
        name: activityUUID
        required: true
        type: string
      - description: Activity
        in: body
        name: activity
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update activity of a dataset for a category
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/activity/{category}/{activityUUID}/delete:
    delete:
      consumes:
      - '*/*'
      description: Delete an activity of a dataset for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Activity UUID
        in: path
        name: activityUUID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an activity of a dataset for a category
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch:
    get:
      consumes:
      - '*/*'
      description: Get all branches of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all branches of a dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}:
    get:
      consumes:
      - '*/*'
      description: Get specific branch of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get specific branch of a dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/delete:
    delete:
      consumes:
      - '*/*'
      description: Delete a branch of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a branch of a dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/hash-status:
    post:
      consumes:
      - '*/*'
      description: Verify dataset hash status to determine if dataset is already uploaded
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Hash value
        in: body
        name: hash
        required: true
        schema:
          $ref: '#/definitions/models.HashRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Verify dataset hash status
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/register:
    post:
      consumes:
      - '*/*'
      description: Register dataset file. Create dataset and default branches if not
        exists
      parameters:
      - description: Dataset file
        in: formData
        name: file
        required: true
        type: file
      - description: Organization UUID
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch name
        in: path
        name: branchName
        required: true
        type: string
      - in: formData
        name: hash
        type: string
      - in: formData
        name: is_empty
        type: boolean
      - in: formData
        name: lineage
        type: string
      - in: formData
        name: storage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Register dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/update:
    post:
      consumes:
      - '*/*'
      description: Update a branch of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a branch of a dataset
      tags:
      - Dataset
...
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}:
    get:
      consumes:
      - '*/*'
      description: Get specific branch version of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get specific branch version of a dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}/log:
    get:
      consumes:
      - '*/*'
      description: Get Log data for dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Log data for dataset
      tags:
      - Dataset
    post:
      consumes:
      - '*/*'
      description: Log data for dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      - description: Data to log
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.LogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Log data for dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/{branchName}/version/{version}/log/{key}:
    get:
      consumes:
      - '*/*'
      description: Get Log data for dataset with specific key
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      - description: Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Log data for dataset with specific key
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/branch/create:
    post:
      consumes:
      - '*/*'
      description: Create a new branch of a dataset
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Data
        in: body
        name: branchName
        required: true
        schema:
          $ref: '#/definitions/models.CreateDatasetBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new branch of a dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/create:
    post:
      consumes:
      - '*/*'
      description: Register dataset file. Create dataset and default branches if not
        exists
      parameters:
      - description: Organization UUID
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Dataset details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateDatasetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create dataset
      tags:
      - Dataset
  /org/{orgId}/dataset/{datasetName}/readme:
    post:
      consumes:
      - '*/*'
      description: Update readme of a dataset for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ReadmeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update readme of a dataset for a category
      tags:
      - Dataset
...
  /org/{orgId}/dataset/{datasetName}/readme/version/{version}:
    get:
      consumes:
      - '*/*'
      description: Get dataset readme
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Dataset Name
        in: path
        name: datasetName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get dataset readme
      tags:
      - Dataset
  /org/{orgId}/dataset/all:
    get:
      consumes:
      - '*/*'
      description: Get all datasets of an organization
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all datasets of an organization
      tags:
      - Dataset
  /org/{orgId}/leave:
    post:
      consumes:
      - '*/*'
      description: Leave organization.
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: User email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Leave organization.
      tags:
      - Organization
  /org/{orgId}/model/{modelName}:
    get:
      consumes:
      - '*/*'
      description: Get specific model of an organization
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Model Name
        in: path
        name: modelName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get specific model of an organization
      tags:
      - Model
...
  /org/{orgId}/model/{modelName}/activity/{category}/{activityUUID}:
    post:
      consumes:
      - '*/*'
      description: Update activity of a model for a category
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Model Name
        in: path
        name: modelName
        required: true
        type: string
      - description: Category
        in: path
        name: category
        required: true
        type: string
      - description: Activity UUID
        in: path
        name: activityUUID
        required: true
        type: string
      - description: Activity
        in: body
        name: activity
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update activity of a model for a category
      tags:
      - Model
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Header for logged in user format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"